FF-116-join-team-modal: add JoinTeamPageMobi and delete ModalJoinTeamMobi, add route to /join-team on btn in /contacts
